requirements.txt
Bootstrap-Flask==1.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-PyMongo==2.3.0
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pymongo==3.9.0
python-dotenv==0.10.3
python-ldap==3.2.0
requests==2.22.0
urllib3==1.25.6
Werkzeug==0.16.0
WTForms==2.2.1
xmltodict==0.12.0
